refactor(jh::concepts, jh::meta, jh::pod, platform, planning): clarify enum and pod semantics
This commit clarifies enum and POD behavior, makes potentially failing operations return expected, and prepares the project for broader refactoring.

jh::concepts:
* Add concepts to identify enum and scoped enum types.

jh::meta:
* Add enum_case and expected.
* Define expected with no-throw behavior; its semantics differ from C++23.

jh::pod:
* Return expected from operations that may fail instead of falling back to nullptr or {}.
* Preserve unchecked array indexing; keep array, pair, and tuple unpacking and get operations compile-time safe.
* Store optional values in a union of T and bytes, allowing direct access as T, T*, or T& without casts.
* Replace hard-coded uint64_t and uint16_t widths, indices, and hash types with size_t; retain uint64_t for 64-bit buffers.
* Make some breaking interface changes that may disrupt the 1.4.x line during the planned broader project refactoring.

platform:
* Remove the 32-bit system restriction in preparation for planned wasm32 support.

planning:
* Plan to redesign the test strategy. Catch2's macros, low-level behavior, and hooks make test behavior difficult to reason about; evaluate switching the test dependency. The dependency has not been migrated in this commit.
* Remove the outdated AGENTS.md; plan to add replacement guidance better suited to this project and AI-assisted development.

+++ examples/example_runtime_arr.cpp
@@ -377,6 +377,7 @@ static EnsureOutput ensure_output_setup;
  */
 
 #include <iostream>
+#include <cstddef>
 #include <iomanip>
 #include <concepts>
 #include <jh/pod>
@@ -1482,7 +1483,7 @@ namespace example {
 
             // --- read values ---
             std::cout << "values:\n";
-            for (std::uint64_t i = 0; i < bits.size(); ++i)
+            for (std::size_t i = 0; i < bits.size(); ++i)
                 std::cout << bits.test(i) << " ";
 
             std::cout << "\n";
@@ -1498,7 +1499,7 @@ namespace example {
             std::cout << "raw words: " << bits.raw_word_count() << "\n";
 
             auto *raw = bits.raw_data();
-            for (std::uint64_t i = 0; i < bits.raw_word_count(); ++i)
+            for (std::size_t i = 0; i < bits.raw_word_count(); ++i)
                 std::cout << "word[" << i << "] = " << raw[i] << "\n";
 
             // --- reset ---
